Move the setup presenters to notices
