+++ app/presenters/notices/setup/ad-hoc-licence.presenter.js
