+++ app/presenters/notices/setup/check.presenter.js
