+++ app/presenters/notices/setup/confirmation.presenter.js
