+++ app/presenters/notices/setup/remove-licences.presenter.js
