+++ app/services/notices/setup/check.service.js
@@ -5,7 +5,7 @@
  * @module CheckService
  */
 
-const CheckPresenter = require('../../../presenters/notifications/setup/check.presenter.js')
+const CheckPresenter = require('../../../presenters/notices/setup/check.presenter.js')
 const DetermineRecipientsService = require('./determine-recipients.service.js')
 const PaginatorPresenter = require('../../../presenters/paginator.presenter.js')
 const RecipientsService = require('./fetch-recipients.service.js')
